requirements.txt
requests
beautifulsoup4
python-Levenshtein
python-whois
fastapi
uvicorn
pydantic